Move and export verbose
