+++ src/api/index.ts
@@ -10,5 +10,6 @@ export * from './pinnedDependencies'
 export * from './locator'
 export * from './modules'
 export * from './table'
+export * from './verbose'
 export { BatchStream } from './typings'
 export { default as logger } from './logger'
