+++ src/lib/sse/index.ts
@@ -4,7 +4,7 @@ import { colossusEndpoint, publicEndpoint } from '../../api/env'
 import { SSEConnectionError } from '../../api/error/errors'
 import { removeVersion } from '../../api/locator'
 import log from '../../api/logger'
-import { isVerbose } from '../../verbose'
+import { isVerbose } from '../../api/verbose'
 import { ErrorKinds } from '../../api/error/ErrorKinds'
 import { ErrorReport } from '../../api/error/ErrorReport'
 import { CustomEventSource } from './CustomEventSource'
