+++ src/modules/rewriter/delete.ts
@@ -5,7 +5,7 @@ import { createInterface } from 'readline'
 import { Rewriter } from '../../api/clients/IOClients/apps/Rewriter'
 import { SessionManager } from '../../api/session/SessionManager'
 import log from '../../api/logger'
-import { isVerbose } from '../../verbose'
+import { isVerbose } from '../../api/verbose'
 import {
   deleteMetainfo,
   handleReadError,
